+++ tests/cli/test_cli.py
@@ -160,6 +160,24 @@ def test_local_evaluate(
 
 
 @pytest.mark.integration
+def test_local_evaluate_demo_resources(
+    test_invalid_config_path: str, test_cli_runner: CliRunner
+) -> None:
+    result = test_cli_runner.invoke(
+        cli,
+        [
+            "--local",
+            "-f",
+            test_invalid_config_path,
+            "evaluate",
+            "demo_resources/",
+        ],
+    )
+    print(result.output)
+    assert result.exit_code == 0
+
+
+@pytest.mark.integration
 def test_evaluate_with_key_pass(
     test_config_path: str, test_cli_runner: CliRunner
 ) -> None:
@@ -245,15 +263,10 @@ def test_evaluate_with_dataset_collection_failed(
 @pytest.mark.parametrize(
     "export_resource", ["system", "dataset", "organization", "datamap"]
 )
-@pytest.mark.parametrize(
-    "manifest_dir",
-    ["demo_resources/", ".fides/"],
-)
 def test_export_resources(
     test_config_path: str,
     test_cli_runner: CliRunner,
     export_resource: str,
-    manifest_dir: str,
 ) -> None:
     """
     Tests that each resource is successfully exported
@@ -266,7 +279,6 @@ def test_export_resources(
             test_config_path,
             "export",
             export_resource,
-            manifest_dir,
             "--dry",
         ],
     )